+++ SharpSource/SharpSource.Test/ThreadSleepInAsyncMethodTests.cs
@@ -409,5 +409,140 @@ namespace ConsoleApplication1
 
             VerifyDiagnostic(original);
         }
+
+        [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/112")]
+        public void ThreadSleepInAsyncMethod_AsyncMethod_Async_ValueTask()
+        {
+            var original = @"
+using System.Threading;
+using System.Threading.Tasks;
+
+class Test
+{
+    async ValueTask MyMethod()
+    {
+        Thread.Sleep(5000);
+    }
+}";
+
+            var result = @"
+using System.Threading;
+using System.Threading.Tasks;
+
+class Test
+{
+    async ValueTask MyMethod()
+    {
+        await Task.Delay(5000);
+    }
+}";
+
+            VerifyDiagnostic(original, "Synchronously sleeping thread in an async method");
+            VerifyFix(original, result);
+        }
+
+        [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/112")]
+        public void ThreadSleepInAsyncMethod_AsyncMethod_Sync_ValueTask()
+        {
+            var original = @"
+using System.Threading;
+using System.Threading.Tasks;
+
+class Test
+{
+    ValueTask MyMethod()
+    {
+        Thread.Sleep(5000);
+        return ValueTask.CompletedTask;
+    }
+}";
+
+            var result = @"
+using System.Threading;
+using System.Threading.Tasks;
+
+class Test
+{
+    ValueTask MyMethod()
+    {
+        Thread.Sleep(5000);
+        return ValueTask.CompletedTask;
+    }
+}";
+
+            VerifyDiagnostic(original, "Synchronously sleeping thread in an async method");
+            VerifyFix(original, result);
+        }
+
+        [TestMethod]
+        public void ThreadSleepInAsyncMethod_AsyncMethod_TopLevel()
+        {
+            var original = @"
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+async Task MyMethod() => Thread.Sleep(5000);
+await MyMethod();
+";
+
+            var result = @"
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+async Task MyMethod() => await Task.Delay(5000);
+await MyMethod();
+";
+
+            VerifyDiagnostic(original, "Synchronously sleeping thread in an async method");
+            VerifyFix(original, result);
+        }
+
+        [TestMethod]
+        public void ThreadSleepInAsyncMethod_AsyncMethod_LocalFunction()
+        {
+            var original = @"
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+async Task Method()
+{
+    async Task MyMethod() => Thread.Sleep(5000);
+    await MyMethod();
+}";
+
+            var result = @"
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+async Task Method()
+{
+    async Task MyMethod() => await Task.Delay(5000);
+    await MyMethod();
+}";
+
+            VerifyDiagnostic(original, "Synchronously sleeping thread in an async method");
+            VerifyFix(original, result);
+        }
+
+        [TestMethod]
+        public void ThreadSleepInAsyncMethod_AsyncMethod_Lambda()
+        {
+            var original = @"
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+async Task MyMethod() {
+	Action lambda = () => Thread.Sleep(32);
+	lambda();
+}
+";
+
+            VerifyDiagnostic(original);
+        }
     }
 }